requirements.txt
amqp==5.1.1
api2pdf==2.1.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
billiard==4.1.0
blinker==1.6.2
Brotli==1.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
clr-loader==0.2.6
cryptography==42.0.5
cssselect2==0.7.0
Flask==3.0.0
Flask-DotEnv==0.1.2
fonttools==4.43.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.2
lxml==4.9.3
MarkupSafe==2.1.3
oscrypto==1.3.0
pdfkit==1.0.0
Pillow==10.0.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.8.0
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
pypdf==4.1.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pythonnet==3.0.3
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.1.0
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.2
urllib3==2.2.1
vine==5.0.0
wcwidth==0.2.8
weasyprint==60.1
webencodings==0.5.1
Werkzeug==3.0.1
zipp==3.16.0
zopfli==0.2.3
